text input: fix problem with autoindenting and ^^D
The following sequence would lead to an abort:

:se ai^Mi^I^M^^Da^D

The problem appears to be that the carat state is overloaded
to track both the current state and something that happened
before ("C_NOCHANGE").  Keep track of "nochange" in a separate
variable.

Problem reported and analyzed by Peter Bex <Peter.Bex@xs4all.nl>

from nvi 1.8x:
Author: Sven Verdoolaege <skimo@kotnet.org>
Date:   Tue Dec 29 19:07:19 2009 +0100