Treat consecutive paragraph indicators as different paragraphs
Consecutive empty lines count toward the same state, so there're 2x states (to get in and out). `^L` and `.PP` are counted as text, hitting those in the text should be treated as getting out of a paragraph and then getting in. From Walter Alejandro Iglesias and Zhihao Yuan in nvi2; ok bluhm@